Modify models to use double type for position.

+++ backend/Data/Models/User.cs
@@ -29,8 +29,8 @@ namespace Data.Models
         public string MobileNumber { get; set; } = null!;
         public string StripeCustomerId { get; set; } = null!;
         public string? StripeAccountId { get; set; } = null!;
-        public decimal? Latitude { get; set; }
-        public decimal? Longitude { get; set; }
+        public double? Latitude { get; set; }
+        public double? Longitude { get; set; }
         public ICollection<Appointment> StudyAppointments { get; set; } = new List<Appointment>();
         public ICollection<Appointment> TutoringAppointments { get; set; } = new List<Appointment>();
         public ICollection<TutoringPost> TutoringPosts { get; set; } = new List<TutoringPost>();
